refactor(New Modules): This commit has cleaned up app.py by migrating some methods to their own files
1. Split out wait_for_file, local_whisper_process & clip_and_sub into their own files
2. Cleaned up import description

Commit has been tested successfully
